+++ Resources/Stamina.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace Resources {
     // These are so empty because the class it inherits, Resource, has all the basic functions
     public class Stamina : Resource {
-        public Stamina(int amount, int maxAmount) : base(amount, maxAmount, "Stamina", (int)ResourceType.Stamina, Color.green) {
+        public Stamina(int amount, int maxAmount) : base(amount, maxAmount, "Stamina", ResourceType.Stamina, Color.green) {
             
         }
